backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "OXN Backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.6",
    "docker>=6.0.1",
    "matplotlib>=3.6.2",
    "numpy>=1.23.5",
    "pandas>=2.1.2",
    "python-dateutil>=2.8.2",
    "PyYAML>=6.0",
    "requests>=2.28.1",
    "schema>=0.7.5",
    "gevent>=22.10.2",
    "locust>=2.14.2",
    "psutil>=5.9.4",
    "pytz>=2022.7.1",
    "python-on-whales>=0.59.0",
    "scipy>=1.10.1",
    "tables>=3.8.0",
    "kubernetes>=28.1.0",
    "jsonschema>=4.19.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "pydantic-to-typescript>=2.0.0",
]

[tool.setuptools]
packages = ["internal"]


[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0"
]
[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_default_fixture_loop_scope = "session"
